Further cleanup.
The biggest change is that the cpsr register
has been moved to its own lesson right before the
cmp instruction is introduced. I feel like this
provides a vary smooth ramp up to the more complex
topics: We introduce the cpsr, which is modified
by cmp, which is used by branching, which is used
for loops and function calls. The we take a step
back to introduce ldr and str, then introduce
the stack, then pop and push.

Aside from that, multiple new comments have been
added.